Fixed silly bug with discretization
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@22961 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 9f8963266f4648a6862ee22a0f0a8cac67b8b6bf

+++ sandbox/distance_field/include/distance_field/voxel_grid.h
@@ -268,7 +268,7 @@ inline void VoxelGrid<T>::setCell(int x, int y, int z, T& obj)
 template<typename T>
 inline int VoxelGrid<T>::getCellFromLocation(Dimension dim, double loc) const
 {
-  return (0.5+loc-origin_[dim])/resolution_[dim];
+  return int(0.5 + (loc-origin_[dim])/resolution_[dim]);
 }
 
 template<typename T>
